+++ packages/core-components/src/layout/ItemCard/ItemCard.tsx
@@ -37,7 +37,7 @@ type ItemCardProps = {
 };
 
 /**
- * This card type has been deprecated. Instead use plain MUI Card and helpers
+ * This card type has been deprecated. Instead use plain Material UI Card and helpers
  * where appropriate.
  *
  *  @example
@@ -57,8 +57,8 @@ type ItemCardProps = {
  *   </Card>
  * ```
  *
- * @deprecated Use plain MUI `<Card>` and composable helpers instead.
- * @see https://material-ui.com/components/cards/
+ * @deprecated Use plain Material UI `<Card>` and composable helpers instead.
+ * @see https://v4.mui.com/components/cards/
  */
 export function ItemCard(props: ItemCardProps) {
   const { description, tags, title, type, subtitle, label, onClick, href } =