+++ packages/core-plugin-api/src/icons/types.ts
@@ -22,8 +22,8 @@ import { ComponentType } from 'react';
  *
  * @remarks
  *
- * The type is based on SvgIcon from MUI, but both do not what the plugin-api
- * package to have a dependency on MUI, nor do we want the props to be as broad
+ * The type is based on SvgIcon from Material UI, but both do not what the plugin-api
+ * package to have a dependency on Material UI, nor do we want the props to be as broad
  * as the SvgIconProps interface.
  *
  * If you have the need to forward additional props from SvgIconProps, you can
@@ -34,11 +34,11 @@ import { ComponentType } from 'react';
  */
 
 export type IconComponent = ComponentType<
-  /* MUI v4 */
+  /* Material UI v4 */
   | {
       fontSize?: 'large' | 'small' | 'default';
     }
-  /* MUI v5: https://mui.com/material-ui/migration/v5-component-changes/#icon */
+  /* Material UI v5: https://mui.com/material-ui/migration/v5-component-changes/#icon */
   | {
       fontSize?: 'medium' | 'large' | 'small';
     }