+++ packages/core-app-api/CHANGELOG.md
@@ -873,7 +873,7 @@
 
   Deprecated the `SamlSession` and `GithubSession` types.
 
-- 784d8078ab: Removed direct and transitive MUI dependencies.
+- 784d8078ab: Removed direct and transitive Material UI dependencies.
 - Updated dependencies
   - @backstage/config@0.1.12
   - @backstage/core-plugin-api@0.5.0
@@ -1072,7 +1072,7 @@
 
   The first and most commonly used one is `createApp` from the new `@backstage/app-defaults` package, which behaves just like the existing `createApp`. In the future this method is likely to be expanded to add more APIs and other pieces into the default setup, for example the Utility APIs from `@backstage/integration-react`.
 
-  The other option that we now provide is to use `createSpecializedApp` from `@backstage/core-app-api`. This is a more low-level API where you need to provide a full set of options, including your own `components`, `icons`, `defaultApis`, and `themes`. The `createSpecializedApp` way of creating an app is particularly useful if you are not using `@backstage/core-components` or MUI, as it allows you to avoid those dependencies completely.
+  The other option that we now provide is to use `createSpecializedApp` from `@backstage/core-app-api`. This is a more low-level API where you need to provide a full set of options, including your own `components`, `icons`, `defaultApis`, and `themes`. The `createSpecializedApp` way of creating an app is particularly useful if you are not using `@backstage/core-components` or Material UI, as it allows you to avoid those dependencies completely.
 
 - 475edb5bc5: move the BehaviorSubject init into the constructor
 - Updated dependencies